config: add global maxOpenFiles setting
Why:

Users need a modern way to configure the open-file limit.

The only rsyslog-side knob was the legacy

directive, which is awkward for RainerScript and YAML configs.

Impact:

Adds global(maxOpenFiles=...) and YAML global.maxOpenFiles.

Existing  configurations remain supported.

Before/After:

Before, only legacy  existed.

After, the same limit can use the modern global object.

Technical Overview:

Register maxOpenFiles as a global configuration parameter.

Store the requested value in the active configuration.

Apply it with setrlimit(RLIMIT_NOFILE) during activation.

Move the shared setrlimit helper to glbl.c.

Keep legacy  immediate behavior unchanged.

Add RainerScript and YAML regression tests.

Document the modern setting and deprecate the legacy directive.

closes https://github.com/rsyslog/rsyslog/issues/4411

With the help of AI-Agents: Codex
